Move ExtensionCatalogProvider files to a separate folder
