+++ packages/suite-base/src/providers/ExtensionCatalogProvider/types.ts
